Data Set Generation
This commit adds code for systematic generation of dataset for the ML
model. The Host/txt and Host/csv files will hold the gestured data.

The naming convention of files in the Host/txt directory is as follows

  <gesture>.txt will hold the raw data corresponding to the <gesture>

The naming convention of files in the Host/txt directory is as follows

  a_b.csv will hold the bth file of the gesture a

Details are outlined in the README.md